Modify switching screens in logged in view through Main
