+++ src/use_case/login/LoginInteractor.java
@@ -2,12 +2,14 @@ package use_case.login;
 
 import entity.User;
 import use_case.login.services.AuthRedirectServerHandler;
+import use_case.login.services.LoginAPIHandler;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
@@ -17,152 +19,156 @@ public class LoginInteractor implements LoginInputBoundary {
     final LoginDataAccessInterface userDataAcessObject;
 
     final LoginOutputBoundary loginPresenter;
+    final LoginAPIHandler handler;
 
-    final String TOKEN_URL = "https://accounts.spotify.com/api/token";
-    final String CLIENT_ID = "1c6992172cb240fd85caa34bae033b94";
-    final String CLIENT_SECRET = "6f843aec369846528a70b621a721e788";
-    static final String REDIRECT_URI = "http://localhost:8888/callback";
+//    final String TOKEN_URL = "https://accounts.spotify.com/api/token";
+//    final String CLIENT_ID = "1c6992172cb240fd85caa34bae033b94";
+//    final String CLIENT_SECRET = "6f843aec369846528a70b621a721e788";
+//    static final String REDIRECT_URI = "http://localhost:8888/callback";
 
-    public LoginInteractor(LoginDataAccessInterface userDataAcessObject, LoginOutputBoundary loginPresenter){
+    public LoginInteractor(LoginDataAccessInterface userDataAcessObject, LoginOutputBoundary loginPresenter, LoginAPIHandler handler){
         this.loginPresenter = loginPresenter;
         this.userDataAcessObject = userDataAcessObject;
-
+        this.handler = handler;
     }
-
-    private String getAccessToken()  {
-        String accessToken = null;
-
-        try {
-            String[] resp = AuthRedirectServerHandler.handleRedirectAndGetResponse(CLIENT_ID, REDIRECT_URI);
-            String authCode = resp[0];
-            String respState = resp[1];
-
-            if (respState != null) {
-                URL url = new URL(TOKEN_URL);
-                HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-                String postData = "code=" + authCode + "&redirect_uri=" + REDIRECT_URI + "&grant_type=authorization_code";
-                byte[] postDataBytes = postData.getBytes(StandardCharsets.UTF_8);
-
-                System.out.println();
-
-                connection.setRequestProperty("Content-Length", Integer.toString(postDataBytes.length));
-
-                connection.setRequestMethod("POST");
-                connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
-
-                String authString = CLIENT_ID + ":" + CLIENT_SECRET;
-                String encodedAuthString = Base64.getEncoder().encodeToString(authString.getBytes());
-                connection.setRequestProperty("Authorization", "Basic " + encodedAuthString);
-
-                connection.setDoOutput(true);
-
-                DataOutputStream wr = new DataOutputStream(connection.getOutputStream());
-                wr.writeBytes(postData);
-
-                int responseCode = connection.getResponseCode();
-                if (responseCode == HttpURLConnection.HTTP_OK) {
-                    BufferedReader in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-                    String inputLine;
-                    StringBuilder response = new StringBuilder();
-
-                    while ((inputLine = in.readLine()) != null) {
-                        response.append(inputLine);
-                    }
-
-                    // Handle the response, which contains the access token and other information
-                    System.out.println("Response: " + response);
-                    accessToken = getJSONDataAccessToken(response.toString());
-
-
-                } else {
-                    // Handle the error
-                    System.err.println("Error response code: " + responseCode);
-                }
-            }
-        }catch(Exception e){
-            System.out.println(e);
-        }
-
-        return accessToken;
-    }
-
-    private static String getJSONDataAccessToken(String responseString) {
-        try {
-            // Locate the "access_token" field
-            int accessTokenIndex = responseString.indexOf("\"access_token\"");
-
-            if (accessTokenIndex != -1) {
-                // Find the opening quote of the access token value
-                int valueStart = responseString.indexOf("\"", accessTokenIndex + 15); // 15 is the length of "access_token" plus the ":"
-
-                if (valueStart != -1) {
-                    // Find the closing quote of the access token value
-                    int valueEnd = responseString.indexOf("\"", valueStart + 1);
-
-                    if (valueEnd != -1) {
-                        // Extract the access token
-                        String accessToken = responseString.substring(valueStart + 1, valueEnd);
-                        return accessToken;
-                    }
-                }
-
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
+//
+//    private String getAccessToken() throws IOException, URISyntaxException, InterruptedException {
+//        String accessToken = null;
+//
+//
+//            String[] resp = AuthRedirectServerHandler.handleRedirectAndGetResponse(CLIENT_ID, REDIRECT_URI);
+//            String authCode = resp[0];
+//            String respState = resp[1];
+//
+//            if (respState != null) {
+//                URL url = new URL(TOKEN_URL);
+//                HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+//                String postData = "code=" + authCode + "&redirect_uri=" + REDIRECT_URI + "&grant_type=authorization_code";
+//                byte[] postDataBytes = postData.getBytes(StandardCharsets.UTF_8);
+//
+//                System.out.println();
+//
+//                connection.setRequestProperty("Content-Length", Integer.toString(postDataBytes.length));
+//
+//                connection.setRequestMethod("POST");
+//                connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+//
+//                String authString = CLIENT_ID + ":" + CLIENT_SECRET;
+//                String encodedAuthString = Base64.getEncoder().encodeToString(authString.getBytes());
+//                connection.setRequestProperty("Authorization", "Basic " + encodedAuthString);
+//
+//                connection.setDoOutput(true);
+//
+//                DataOutputStream wr = new DataOutputStream(connection.getOutputStream());
+//                wr.writeBytes(postData);
+//
+//                int responseCode = connection.getResponseCode();
+//                if (responseCode == HttpURLConnection.HTTP_OK) {
+//                    BufferedReader in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+//                    String inputLine;
+//                    StringBuilder response = new StringBuilder();
+//
+//                    while ((inputLine = in.readLine()) != null) {
+//                        response.append(inputLine);
+//                    }
+//
+//                    // Handle the response, which contains the access token and other information
+//                    System.out.println("Response: " + response);
+//                    accessToken = getJSONDataAccessToken(response.toString());
+//
+//
+//                } else {
+//                    // Handle the error
+//                    System.err.println("Error response code: " + responseCode);
+//                }
+//            }
+//
+//
+//        return accessToken;
+//    }
+//
+//    private static String getJSONDataAccessToken(String responseString) {
+//        try {
+//            // Locate the "access_token" field
+//            int accessTokenIndex = responseString.indexOf("\"access_token\"");
+//
+//            if (accessTokenIndex != -1) {
+//                // Find the opening quote of the access token value
+//                int valueStart = responseString.indexOf("\"", accessTokenIndex + 15); // 15 is the length of "access_token" plus the ":"
+//
+//                if (valueStart != -1) {
+//                    // Find the closing quote of the access token value
+//                    int valueEnd = responseString.indexOf("\"", valueStart + 1);
+//
+//                    if (valueEnd != -1) {
+//                        // Extract the access token
+//                        String accessToken = responseString.substring(valueStart + 1, valueEnd);
+//                        return accessToken;
+//                    }
+//                }
+//
+//            }
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//        return null;
+//    }
     //@Override
     public void execute() throws IOException {
-        String accessToken = getAccessToken();
+//        String accessToken = getAccessToken();
+//
+//        // Create the URL for the Spotify API endpoint
+//        String endpoint = "https://api.spotify.com/v1/me";
+//        URL url = new URL(endpoint);
+//
+//        // Open a connection to the URL
+//        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+//
+//        // Set the request method to GET
+//        connection.setRequestMethod("GET");
+//
+//        // Set the Authorization header with the access token
+//        connection.setRequestProperty("Authorization", "Bearer " + accessToken);
+//
+//        // Get the response code
+//        int responseCode = connection.getResponseCode();
+//        BufferedReader in;
+//        if (responseCode == 200) {
+//            // The request was successful (HTTP status code 200)
+//
+//            // Read the response
+//            in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+//        }
+//        else {
+//            in = new BufferedReader(new InputStreamReader(connection.getErrorStream()));
+//        }
+//        String inputLine;
+//        String[] splitLine;
+//        ArrayList<String[]> response = new ArrayList<>();
+//        StringBuilder stringRepresentation = new StringBuilder();
+//        while ((inputLine = in.readLine()) != null) {
+//            stringRepresentation.append(inputLine);
+//            splitLine = inputLine.split(";");
+//            response.add(splitLine);
+//        }
+//        in.close();
+//        // Print the response (JSON data)
+//        System.out.println("Response: " + stringRepresentation);
+//
+//        // Close the connection
+//        connection.disconnect();
 
-        // Create the URL for the Spotify API endpoint
-        String endpoint = "https://api.spotify.com/v1/me";
-        URL url = new URL(endpoint);
+        try {
+            User user = this.handler.getLoginUserInfo();
+            userDataAcessObject.loginUser(user);
 
-        // Open a connection to the URL
-        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-        // Set the request method to GET
-        connection.setRequestMethod("GET");
-
-        // Set the Authorization header with the access token
-        connection.setRequestProperty("Authorization", "Bearer " + accessToken);
-
-        // Get the response code
-        int responseCode = connection.getResponseCode();
-        BufferedReader in;
-        if (responseCode == 200) {
-            // The request was successful (HTTP status code 200)
-
-            // Read the response
-            in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+            LoginOutputData loginOutputData = new LoginOutputData(user.getName(), false);
+            loginPresenter.prepareSuccessView(loginOutputData);
+        }catch (Exception e){
+            LoginOutputData loginOutputData = new LoginOutputData("", true);
+            loginPresenter.prepareFailView(e.getMessage());
         }
-        else {
-            in = new BufferedReader(new InputStreamReader(connection.getErrorStream()));
-        }
-        String inputLine;
-        String[] splitLine;
-        ArrayList<String[]> response = new ArrayList<>();
-        StringBuilder stringRepresentation = new StringBuilder();
-        while ((inputLine = in.readLine()) != null) {
-            stringRepresentation.append(inputLine);
-            splitLine = inputLine.split(";");
-            response.add(splitLine);
-        }
-        in.close();
-        // Print the response (JSON data)
-        System.out.println("Response: " + stringRepresentation);
 
-        // Close the connection
-        connection.disconnect();
-        HashMap<String, String> userData = stringToMap(stringRepresentation);
-        User user = new User(userData.get("display_name"), accessToken);
-
-        userDataAcessObject.loginUser(user);
-
-        LoginOutputData loginOutputData = new LoginOutputData(user.getName(), false);
-        loginPresenter.prepareSuccessView(loginOutputData);
     }
 
     private static HashMap<String, String> stringToMap(StringBuilder response){
